Move move-page views to wagtail.admin.views.pages.move
